Move ALL THE CLASSES
* Turns out we cannot have subdirectories, which is lame, but manageable
* It may be beneficial to pull out the RTC stuff into a single library
  with a separate repo or whatever if necessary. That way multiple chips
  can be supported
* I hope the maintainers like this refactor. If not, I may just be working
  on my own fork for Watchy...
