+++ src/WatchyRTC.cpp
@@ -2,30 +2,8 @@
 
 WatchyRTC::WatchyRTC() {}
 
-// TODO: We can probably put all of this logic into AbstractRTC as a class
-// function. It would simplify this class even more, which would be nice
-bool WatchyRTC::_canConnectTo(int addr) {
-    byte error;
-    Wire.beginTransmission(addr);
-    error = Wire.endTransmission(addr);
-    return error == 0;
-}
-
 void WatchyRTC::init(){
-    if (_canConnectTo(RTC_DS_ADDR)) {
-        rtcType = DS3232_RTC_TYPE;
-        _rtc = new DS3232();
-        return;
-    }
-
-    if (_canConnectTo(RTC_PCF_ADDR)) {
-        rtcType = PCF8563_RTC_TYPE;
-        _rtc = new PCF8563();
-        return;
-    }
-
-    rtcType = NO_RTC_TYPE;
-    _rtc = new AbstractRTC();
+    _rtc = AbstractRTC::init();
 }
 
 void WatchyRTC::config(String datetime){
@@ -46,4 +24,8 @@ void WatchyRTC::set(tmElements_t tm){
 
 uint8_t WatchyRTC::temperature(){
     return _rtc->temperature();
+}
+
+int WatchyRTC::rtcType() {
+    return _rtc->rtcType();
 }
